+++ libraries/stdlib/js/src/kotlin/js/cosh.kt
