+++ libraries/stdlib/js/src/kotlin/js/expm1.kt
@@ -1,12 +1,10 @@
 /*
- * Copyright 2010-2021 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Copyright 2010-2022 JetBrains s.r.o. and Kotlin Programming Language contributors.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
 @file:JsQualifier("Math")
 package kotlin.js
-
-import kotlin.js.math.defineTaylorNBound
 
 @PublishedApi
 @JsName("expm1")
