+++ libraries/stdlib/js/src/kotlin/js/log10.kt
