+++ libraries/stdlib/js/src/kotlin/js/log2.kt
