+++ libraries/stdlib/js/src/kotlin/js/trunc.kt
